Project was changed according to the comment of the reviewer
