+++ .idea/misc.xml
@@ -13,7 +13,7 @@
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout-v28/activity_player.xml" value="0.3078125" />
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/activity_player.xml" value="0.25" />
-        <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/activity_search.xml" value="0.1" />
+        <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/activity_search.xml" value="0.25" />
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/activity_settings.xml" value="0.32228915662650603" />
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/player.xml" value="0.33" />
         <entry key="..\:/Android/AndroidStudioProjects/PlaylistMaker/app/src/main/res/layout/track_card.xml" value="0.33" />
