+++ app/src/main/java/com/practicum/playlistmaker/TrackAdapter.kt
@@ -1,5 +1,7 @@
 package com.practicum.playlistmaker
 
+import android.os.Handler
+import android.os.Looper
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
@@ -17,7 +19,7 @@ class TrackAdapter(var track: List<Track>) : RecyclerView.Adapter<TrackViewHolde
     override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
         holder.bind(track[position])
         holder.itemView.setOnClickListener { //слушатель нажатия на трэк
-            if (SearchActivity.clickDebounce()) {
+            if (clickDebounce()) {
                 searchHistory.addTrackInHistory(track[position]) //функция добавления трэка
                 SearchActivity.startActivity(track[position], it.context) //вызываем новую активити
             }
@@ -26,5 +28,22 @@ class TrackAdapter(var track: List<Track>) : RecyclerView.Adapter<TrackViewHolde
 
     override fun getItemCount(): Int {
         return track.size
+    }
+
+    private var isClickAllowed = true
+
+    private val handler = Handler(Looper.getMainLooper())
+
+    fun clickDebounce() : Boolean { // ограничение нажатия на элементы списка не чаще одного раза в секунду
+        val current = isClickAllowed
+        if (isClickAllowed) {
+            isClickAllowed = false
+            handler.postDelayed({ isClickAllowed = true }, CLICK_DEBOUNCE_DELAY)
+        }
+        return current
+    }
+
+    companion object {
+        private const val CLICK_DEBOUNCE_DELAY = 1000L // ограничение нажатия на элементы списка не чаще одного раза в секунду
     }
 }
