+++ app/src/main/res/layout/activity_search.xml
@@ -50,14 +50,12 @@
             android:src="@drawable/ic_clear_button" />
     </FrameLayout>
 
-    <FrameLayout android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
     <ProgressBar
         android:id="@+id/progressBar"
         android:layout_gravity="center"
-        android:layout_width="@dimen/progress_bar_size"
-        android:layout_height="@dimen/progress_bar_size"
+        android:layout_width="@dimen/progressbar_size"
+        android:layout_height="@dimen/progressbar_size"
+        android:layout_marginTop="@dimen/margin_top_progressbar"
         android:indeterminateTint="@color/blue"
         android:visibility="gone" />
 
@@ -127,6 +125,4 @@
             style="@style/ButtonUpdate"
             android:text="@string/clear_history_button" />
     </LinearLayout>
-    </FrameLayout>
-
 </LinearLayout>
